+++ tests/unit/utils/skos/test_codelist.py
@@ -66,12 +66,12 @@ def test_get_codelist_col_title_by_property_url_notation():
         csvw_metadata_rdf_graph, csv_url
     )
     result_primary_key_col_names_by_csv_url = (
-        csvw_inspector.get_codelist_primary_key_by_csv_url()
+        csvw_inspector.get_table_schema_properties().table_schema_properties
     )
 
     unique_identifier = get_codelist_col_title_from_col_name(
         result_code_list_cols.columns,
-        result_primary_key_col_names_by_csv_url.primary_key_col_names[0].value,
+        result_primary_key_col_names_by_csv_url[0].primary_key_col_names,
     )
 
     assert unique_identifier == "Notation"
